LayoutTests/media/track/track-text-track-cue-list-expected.txt

Tests TextTrackCueList functionality: length, operator[], and getCueById()


*** Testing TextTrackCueList length
EXPECTED (cues.length == '4') OK

*** Testing TextTrackCueList [] operator
EXPECTED (cues[0].id == '1') OK
EXPECTED (cues[3].id == '4') OK
EXPECTED (cues[4] == 'undefined') OK

*** Testing TextTrackCueList getCueById()
EXPECTED (cues.getCueById('1').startTime == '0') OK
EXPECTED (cues.getCueById('4').startTime == '121') OK
EXPECTED (cues.getCueById('junk') == 'undefined') OK
END OF TEST
